requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.27.0,<0.28
paramiko~=3.4.0
pandas~=2.2.2
sshtunnel~=0.4.0
schedule>=1.2.2,<1.3
channels>=4.0.0,<4.1
daphne>=4.0.0,<4.1